Function synthrs::sample::samples_from_wave_bytes

pub fn samples_from_wave_bytes(bytes: Vec<u8>) -> Result<(Vec<f64>, usize)>

Given a bunch of bytes for a wave file, extract a Vec<f64> of samples from it and the size of that vec

use std::fs::OpenOptions;
use std::io::{BufReader, Read};
use std::path::Path;
use synthrs::sample::samples_from_wave_bytes;
use synthrs::writer::read_wav_file;

let path = Path::new("./tests/assets/sine.wav");
let file = OpenOptions::new().read(true).open(&path).unwrap();
let mut reader = BufReader::new(file);
let mut buf: Vec<u8> = Vec::new();
reader.read_to_end(&mut buf);

let (samples, num_samples) = samples_from_wave_bytes(buf).unwrap();