Module synthrs::synthesizer

Generate audio samples from code.

The following code generates a 0.1s long, 16-bit, 440Hz sine_wave at a 44100Hz sample rate. It then writes the generated samples into a 44100Hz WAV file at out/sine.wav.

use synthrs::wave::sine_wave;
use synthrs::writer::write_wav_file;
use synthrs::synthesizer::{quantize_samples, make_samples};

write_wav_file(
    "out/sine.wav",
    44_100,
    &quantize_samples::<i16>(&make_samples(0.1, 44_100, sine_wave(440.0))),
).expect("failed to write wav");

See: examples/simple.rs

Structs

SamplesIter
An iterator that generates samples of a waveform at a given sample rate

Functions

generate
Invokes the waveform function f at time t to return the amplitude at that time.

make_samples
Given a generator waveform, returns a Vec<f64> of raw samples (not normalised or quantised)

make_samples_from_midi
Generates samples from a MIDI file. Supports only one instrument. Instrument can be any generator.

make_samples_from_midi_file
Generates samples from a MIDI file

peak_normalize
Peak normalizes a Vec<f64> of samples such that the maximum and minimum amplitudes of the Vec<f64> samples are within the range [-1.0, 1.0]

quantize
Quantizes a f64 sample into T. Convert from [-1.0f64, 1.0] to take up full quantization range of type T. Behaviour is not guaranteed for T where T is unsigned.

quantize_samples
Quantizes a Vec<f64> of samples into Vec<T>.

unquantize
Reverses a quantization from T into f64. Convert from T to take up full quantization range of f64.

unquantize_samples
Reverses quantization of Vec<T> into a Vec<f64>.