Function synthrs::wave::rising_linear

pub fn rising_linear(
    start_frequency: f64, 
    end_frequency: f64, 
    increment_per_sample: f64
) -> impl Fn(f64) -> f64

rising_linear is a stateful generator function. Starting from start_frequency, it increases the output frequency by increment_per_sample each time it is called, and loops back to start_frequency when it is above end_frequency.

This is mainly an example on how to do stateful generator functions. This is achieved using interior mutability. See the source for details on how this is achieved.