Function synthrs::wave::sampler

pub fn sampler(
    frequency: f64, 
    samples: *const Vec<f64>, 
    sample_length: usize, 
    sample_frequency: f64, 
    sample_rate: usize
) -> impl Fn(f64) -> f64

sampler creates a a generator function given a bunch of samples. Different frequencies are generated using a simple pitch shift. This function is unsafe because we want to share sample without resorting to a 'static lifetime (making it easier to use with bindgen).

frequency: The frequency passed in to the generator

samples: The raw pointer to the samples to be used

sample_length: Length of samples

sample_frequency: The frequency of the sample provided. This is used to calculate how much to shift the pitch.

sample_rate: The sample rate of the given sample

use synthrs::sampler;
use synthrs::wave;

let (piano_sample, sample_length) =
       sample::samples_from_wave_file("examples/assets/piano110hz.wav").unwrap();

let frequency_to_generate = 110.0;
let sampler = wave::sampler(frequency, &piano_sample, sample_length, 110.0, 44_100)