Module synthrs::writer

Writes audio samples to files. (Reads too, this module needs a rename)

Audacity PCM import settings

File > Import > Raw Data...

Signed 16-bit PCM

Little-endian

1 Channel (Mono)

Sample rate: 44_100Hz (or whatever your samples generated have)

Structs

Wave
Representation of a WAV file. Does not contain fields for extended WAV formats.

Functions

read_wav
Reads a wave file. Only supports mono 16-bit, little-endian, signed PCM WAV files

read_wav_file
Reads a wave file given a file path. Convenience wrapper around crate::writer::read_wav_file.

write_pcm
Writes a bunch of &[i16] samples to a Write as raw PCM. See module documentation for PCM settings.

write_pcm_file
Creates a file at filename and writes a bunch of &[i16] samples to it as a PCM file. See module documentation for PCM settings.

write_wav
Writes a bunch of &[i16] samples to a Write. Also see synthrs::writer::write_wav_file.

write_wav_file
Creates a file at filename and writes a bunch of &[i16] samples to it as a WAVE file