Function synthrs::wave::karplus_strong

pub fn karplus_strong<F: Fn(f64) -> f64>(
    generator: F, 
    attack: f64, 
    decay: f64, 
    sharpness: f64, 
    sample_rate: f64
) -> impl Fn(f64) -> f64

Bastardised and butchered generic Karplus-Strong synthesis. Try a Sawtooth, or even a Bell wave.

This is an example of a generator function using another generator function. In this case, karplus_strong wraps around a generator function and applies a poor emulation of a real-world object over it.

attack in seconds decay in seconds sharpness 0-1 is decent sample_rate in hertz (eg, 44_100.0)

use synthrs::wave;

let karplus_sawtooth_generator =
    |frequency: f64| {
        wave::karplus_strong(wave::sawtooth_wave(frequency), 0.01, 1.0, 0.9, 44_100.0)
    };